requirements.txt
anthropic>=0.82.0
Pillow>=12.0.0
opencv-python>=4.10.0
numpy>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pyyaml>=6.0
python-dotenv>=1.0.0
jinja2>=3.1.0
aiofiles>=24.0.0
